Use GFP_ATOMIC inside the irq locked code
This code is inside an irq lock, which means that we get a BUG warning
for calling kzalloc/kmalloc when allowed to sleep.
Change to GFP_ATOMIC to tell the kernel we really want this memory
even though this code is inside a lock.

Also, swap the print_rule code to not need to allocate