Fix out of bounds memmove
When configured with the same map4to6 / map6to4 prefixes traffic gets
looped by the tunnel. This can lead to the ipv6 payload of the generated
ICMP error message being too small, and the adjustment that the nat46
module makes trying to move a negative length data block.

If the v6 length is too small, just leave the payload as is.